[Wallet]: Migrate NFTs View to Screens Directory
