+++ components/brave_wallet_ui/components/desktop/popup-modals/nft_more_popup/nft_more_popup.tsx
@@ -6,10 +6,10 @@
 import * as React from 'react'
 import Icon from '@brave/leo/react/icon'
 
-import { getLocale } from '../../../../../../../common/locale'
+import { getLocale } from '../../../../../common/locale'
 
 // Styled Components
-import { ButtonMenu } from '../../../../wallet-menus/wellet-menus.style'
+import { ButtonMenu } from '../../wallet-menus/wellet-menus.style'
 
 interface Props {
   children: React.ReactNode
