+++ components/brave_wallet_ui/page/router/unlocked_wallet_routes.tsx
@@ -38,7 +38,7 @@ import { AccountSettingsModal } from '../../components/desktop/popup-modals/acco
 import {
   WalletPageWrapper, //
 } from '../../components/desktop/wallet-page-wrapper/wallet-page-wrapper'
-import { NftCollection } from '../../components/desktop/views/nfts/components/nft_collection'
+import { NftCollection } from '../screens/nfts/nft_collection'
 import { Banners } from '../../components/desktop/banners/banners'
 import {
   BackupWalletRoutes, //
