+++ components/brave_wallet_ui/page/screens/nfts/components/auto_discovery_empty_state/auto_discovery_empty_state.styles.tsx
@@ -2,11 +2,12 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import styled from 'styled-components'
 import * as leo from '@brave/leo/tokens/css/variables'
 import Ring from '@brave/leo/react/progressRing'
 
-import { WalletButton } from '../../../../../shared/style'
+import { WalletButton } from '../../../../../components/shared/style'
 
 export const StyledWrapper = styled.div`
   display: flex;
