+++ components/brave_wallet_ui/page/screens/nfts/components/auto_discovery_empty_state/auto_discovery_empty_state.tsx
@@ -2,6 +2,7 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
 
 // utils
@@ -13,8 +14,12 @@ import {
   Heading,
   LoadingRing,
   StyledWrapper,
-} from './auto-discovery-empty-state.styles'
-import { Text, Row, VerticalSpace } from '../../../../../shared/style'
+} from './auto_discovery_empty_state.styles'
+import {
+  Text,
+  Row,
+  VerticalSpace,
+} from '../../../../../components/shared/style'
 
 interface Props {
   isRefreshingTokens: boolean
