+++ components/brave_wallet_ui/page/screens/nfts/components/nft_grid_views/nft_collection_grid_view_item/nft_collection_grid_view_item.styles.ts
