+++ components/brave_wallet_ui/page/screens/nfts/components/nft_grid_views/nft_grid_view_item/nft_grid_view_item.stories.tsx
@@ -10,7 +10,7 @@ import {
   WalletPanelStory, //
 } from '../../../../../../stories/wrappers/wallet-panel-story-wrapper'
 import { PanelWrapper } from '../../../../../../panel/panel_wrapper/panel_wrapper'
-import { NFTGridViewItem } from './nft-grid-view-item'
+import { NFTGridViewItem } from './nft_grid_view_item'
 
 // mocks
 import {
