+++ components/brave_wallet_ui/page/screens/nfts/components/nft_grid_views/nft_grid_view_item_skeleton/nft_grid_view_item_skeleton.stories.tsx
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import { NftGridViewItemSkeleton } from './nft-grid-view-item-skeleton'
+import { NftGridViewItemSkeleton } from './nft_grid_view_item_skeleton'
 
 export const _NftGridViewItemSkeleton = {}
 export default {
