+++ components/brave_wallet_ui/page/screens/nfts/components/nft_grid_views/nft_grid_view_item_skeleton/nft_grid_view_item_skeleton.tsx
@@ -2,10 +2,13 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
-import { LoadingSkeleton } from '../../../../../shared/loading-skeleton/index'
-import { VerticalSpace } from '../../../../../shared/style'
-import { NFTWrapper } from './style'
+import {
+  LoadingSkeleton, //
+} from '../../../../../../components/shared/loading-skeleton/index'
+import { VerticalSpace } from '../../../../../../components/shared/style'
+import { NFTWrapper } from '../nft_grid_views.style'
 
 export const NftGridViewItemSkeleton = () => {
   return (
