+++ components/brave_wallet_ui/page/screens/nfts/components/nft_group_selector/nft_group_selector.styles.ts
@@ -2,11 +2,12 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import styled from 'styled-components'
 import Icon from '@brave/leo/react/icon'
 import * as leo from '@brave/leo/tokens/css/variables'
 
-import { WalletButton, Text } from '../../../../../shared/style'
+import { WalletButton, Text } from '../../../../../components/shared/style'
 
 export const DropdownContainer = styled.div`
   position: relative;
