+++ components/brave_wallet_ui/page/screens/nfts/components/nfts_empty_state/nfts_empty_state.style.ts
@@ -2,10 +2,11 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import * as leo from '@brave/leo/tokens/css/variables'
 import styled from 'styled-components'
 
-import { WalletButton } from '../../../../../shared/style'
+import { WalletButton } from '../../../../../components/shared/style'
 
 export const StyledWrapper = styled.div`
   display: flex;
