+++ components/brave_wallet_ui/page/screens/nfts/nfts.styles.tsx
@@ -1,16 +1,16 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as leo from '@brave/leo/tokens/css/variables'
 import styled from 'styled-components'
 
 // Shared Styles
-import { Row, ScrollableColumn } from '../../../../shared/style'
+import { Row, ScrollableColumn } from '../../../components/shared/style'
 import {
   layoutPanelWidth, //
-} from '../../../wallet-page-wrapper/wallet-page-wrapper.style'
+} from '../../../components/desktop/wallet-page-wrapper/wallet-page-wrapper.style'
 
 export const NftGrid = styled.div<{
   padding?: string
