+++ components/brave_wallet_ui/page/screens/nfts/nfts.tsx
@@ -1,53 +1,51 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
 import { useHistory } from 'react-router-dom'
 import { skipToken } from '@reduxjs/toolkit/query'
 
 // redux
-import { useAppDispatch } from '../../../../../common/hooks/use-redux'
+import { useAppDispatch } from '../../../common/hooks/use-redux'
 
 // types
-import {
-  BraveWallet,
-  NftDropdownOptionId,
-} from '../../../../../constants/types'
+import { BraveWallet, NftDropdownOptionId } from '../../../constants/types'
 
 // hooks
-import { useAccountsQuery } from '../../../../../common/slices/api.slice.extra'
+import { useAccountsQuery } from '../../../common/slices/api.slice.extra'
 import {
   useBalancesFetcher, //
-} from '../../../../../common/hooks/use-balances-fetcher'
+} from '../../../common/hooks/use-balances-fetcher'
 import {
   useLocalStorage,
   useSyncedLocalStorage, //
-} from '../../../../../common/hooks/use_local_storage'
+} from '../../../common/hooks/use_local_storage'
 
 // selectors
 import {
   useSafeUISelector,
   useSafeWalletSelector,
-} from '../../../../../common/hooks/use-safe-selector'
-import { UISelectors, WalletSelectors } from '../../../../../common/selectors'
+} from '../../../common/hooks/use-safe-selector'
+import { UISelectors, WalletSelectors } from '../../../common/selectors'
 
 // actions
-import { WalletActions } from '../../../../../common/actions'
-import { WalletPageActions } from '../../../../../page/actions'
+import { WalletActions } from '../../../common/actions'
+import { WalletPageActions } from '../../actions'
 
 // utils
 import { getLocale } from '$web-common/locale'
 import {
   LOCAL_STORAGE_KEYS, //
-} from '../../../../../common/constants/local-storage-keys'
+} from '../../../common/constants/local-storage-keys'
 import {
   useGetNftAssetIdsByCollectionRegistryQuery,
   useGetNftDiscoveryEnabledStatusQuery,
   useGetSimpleHashSpamNftsQuery,
   useGetUserTokensRegistryQuery,
   useSetNftDiscoveryEnabledMutation,
-} from '../../../../../common/slices/api.slice'
+} from '../../../common/slices/api.slice'
 import {
   compareTokensByName,
   filterTokensByNetworks,
@@ -59,51 +57,61 @@ import {
   isTokenWatchOnly,
   searchNftCollectionsAndGetTotalNftsFound,
   searchNfts,
-} from '../../../../../utils/asset-utils'
-import { useQuery } from '../../../../../common/hooks/use-query'
+} from '../../../utils/asset-utils'
+import { useQuery } from '../../../common/hooks/use-query'
 import {
   makePortfolioAssetRoute,
   makePortfolioNftCollectionRoute,
   makePortfolioNftsRoute,
-} from '../../../../../utils/routes-utils'
+} from '../../../utils/routes-utils'
 import {
   selectAllVisibleUserNFTsFromQueryResult,
   selectHiddenNftsFromQueryResult, //
-} from '../../../../../common/slices/entities/blockchain-token.entity'
+} from '../../../common/slices/entities/blockchain-token.entity'
 import {
   getLastPageNumber,
   getListPageItems,
-} from '../../../../../utils/pagination_utils'
+} from '../../../utils/pagination_utils'
 
 // components
-import SearchBar from '../../../../shared/search-bar'
-import { NFTGridViewItem } from '../../portfolio/components/nft-grid-view/nft-grid-view-item'
-import { EnableNftDiscoveryModal } from '../../../popup-modals/enable-nft-discovery-modal/enable-nft-discovery-modal'
-import { AutoDiscoveryEmptyState } from './auto-discovery-empty-state/auto-discovery-empty-state'
+import SearchBar from '../../../components/shared/search-bar'
+import {
+  NFTGridViewItem, //
+} from './components/nft_grid_views/nft_grid_view_item/nft_grid_view_item'
+import {
+  EnableNftDiscoveryModal, //
+} from '../../../components/desktop/popup-modals/enable-nft-discovery-modal/enable-nft-discovery-modal'
+import {
+  AutoDiscoveryEmptyState, //
+} from './components/auto_discovery_empty_state/auto_discovery_empty_state'
 import {
   NftGridViewItemSkeleton, //
-} from '../../portfolio/components/nft-grid-view/nft-grid-view-item-skeleton'
-import { Pagination } from '../../../../shared/pagination/pagination'
+} from './components/nft_grid_views/nft_grid_view_item_skeleton/nft_grid_view_item_skeleton'
+import { Pagination } from '../../../components/shared/pagination/pagination'
 import {
   NftDropdown,
   NftDropdownOption,
-} from './nft-group-selector/nft-group-selector'
+} from './components/nft_group_selector/nft_group_selector'
 import {
   NftCollectionGridViewItem, //
-} from '../../portfolio/components/nft-grid-view/nft-collection-grid-view-item'
+} from './components/nft_grid_views/nft_collection_grid_view_item/nft_collection_grid_view_item'
 
 // styles
 import { NFTListWrapper, NftGrid } from './nfts.styles'
-import { Column, Row } from '../../../../shared/style'
-import { AddOrEditNftModal } from '../../../popup-modals/add-edit-nft-modal/add-edit-nft-modal'
-import { NftsEmptyState } from './nfts-empty-state/nfts-empty-state'
+import { Column, Row } from '../../../components/shared/style'
+import {
+  AddOrEditNftModal, //
+} from '../../../components/desktop/popup-modals/add-edit-nft-modal/add-edit-nft-modal'
+import {
+  NftsEmptyState, //
+} from './components/nfts_empty_state/nfts_empty_state'
 import {
   ButtonIcon,
   PortfolioActionButton,
   SearchBarWrapper,
   ControlBarWrapper,
   ContentWrapper,
-} from '../../portfolio/style'
+} from '../page-screen.styles'
 
 interface Props {
   onShowPortfolioSettings?: () => void
